beamer: tests: fix tests after move
